implement grid view
I initially tried to use a table to implement the grid view because I wanted to challenge myself, as it seemed like using a grid would be easier. However, it required me to nest two loops and needed some extra elements. In the end, I decided to go with a grid.