implement drawLine feature
This might have been the most challenging part of this section. I initially searched online for 'mousedown,' 'mouseover,' and 'mouseup' events, and then had to figure out how to display the selected color on the current cell. I couldn't quite get it right at first. To better understand what was happening, I used a lot of 'console.log' statements to debug. While trying to manage it with the 'drawingInProgress' variable, the drawing would stop when I set 'drawingInProgress' to false on the 'mouseleave' event. To solve this issue, I created a variable called 'cellClicked,' and suddenly my problem was resolved.