add darkness to color
I found this situation a bit challenging, and eventually, I added a 'count' attribute to the cells. To be honest, I wasn't entirely comfortable with this approach because I wasn't sure if it was the right one, but it solved my problem. In the 'addDarknessToColor' function, I increment the 'count' each time it runs, and then I subtract the desired darkness amount multiplied by the 'count' from the color value.

Initially, the darkening effect only worked when I hovered the mouse over the cell, but it didn't work when I clicked on the cell. To address this, I added an additional 'paintCell' function to the mousedown event.